% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_counts.R
\name{aa_counts}
\alias{aa_counts}
\title{aa_counts}
\usage{
aa_counts(peptide)
}
\arguments{
\item{peptide}{is a string containing a sequence of amino acids constituting a peptide}
}
\value{
outputs a bar-plot with counts of all occuring amino acids in the input peptide
}
\description{
aa_counts
}
