% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_frequency_plot.R
\name{aa_frequency_plot}
\alias{aa_frequency_plot}
\title{Create a barplot representing the frequencies of each aminoacids  in the given protein}
\usage{
aa_frequency_plot(protein)
}
\arguments{
\item{protein}{the protein of which to count the frequencies of the aminoacids}
}
\value{
Barplot of the aa frequencies
}
\description{
Create a barplot representing the frequencies of each aminoacids  in the given protein
}
\examples{
aa_frequency_plot("ATTTTWWWHH")
}
