% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_num_plot.R
\name{AA_num_plot}
\alias{AA_num_plot}
\title{Visualization of the number of each unique amino acid}
\usage{
AA_num_plot(AA_seq)
}
\arguments{
\item{AA_seq}{This is the translated amino acid sequence}
}
\value{
The function returns the bar plot
}
\description{
Visualization of the number of each unique amino acid
}
\examples{
AAplot <- AA_num_plot("IPTAAKNN")

}
