% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codon_code}
\alias{codon_code}
\title{codon_code}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codon_code
}
\description{
A named list where the each codon (three nucleotide sequence)
is matched with the code amminoacid
}
\keyword{datasets}
