% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonstring.R
\name{codon_string}
\alias{codon_string}
\title{Translate Codon Sequences to Amino Acids
This function takes a vector of codon sequences and returns a single string
of corresponding amino acids by using a predefined table (`codon_code`).}
\usage{
codon_string(codons)
}
\arguments{
\item{codons}{A character vector of codon sequences, where each sequence should be a valid codon (e.g., "UUU", "UCU").}
}
\value{
A single string that represents the sequence of amino acids
corresponding to the provided codons.
}
\description{
Translate Codon Sequences to Amino Acids
This function takes a vector of codon sequences and returns a single string
of corresponding amino acids by using a predefined table (`codon_code`).
}
\examples{
codon_code <- c("UUU" = "F", "UCU" = "S", "UAU" = "Y", "UGU" = "C",
                "UUC" = "F", "UCC" = "S", "UAC" = "Y", "UGC" = "C")
codon_string(c("UUU", "UCU", "UGC"))  # Returns "FSC"

}
