% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codons}
\alias{codons}
\title{A standard codon table.}
\format{

}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codons
}
\description{
This table contains all the codons and their corresponding amino acids.
}
\keyword{datasets}
