% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_generator.R
\name{dna_generator}
\alias{dna_generator}
\title{DNA Generator}
\usage{
dna_generator(n)
}
\arguments{
\item{n}{where n is the size of the DNA strand}
}
\value{
returns the DNA strand of the desired size as a string without spaces
}
\description{
DNA Generator
}
\examples{
dna_generator(10)
}
