% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codons.R
\name{get_codons}
\alias{get_codons}
\title{Extract codons from a sequence}
\usage{
get_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{input sequence to parse into the codons}

\item{start}{start position for the codon extraction}
}
\value{
list of codons
}
\description{
Extract codons from a sequence
}
\examples{
get_codons("GUCCCAAGUC")
}
