% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_DNA.R
\name{Make_DNA}
\alias{Make_DNA}
\title{Title: Makes DNA of a given len}
\usage{
Make_DNA(len_DNA)
}
\arguments{
\item{len_DNA}{the length of the DNA string}
}
\value{
The function returens a dna string whit a given length
}
\description{
Title: Makes DNA of a given len
}
\examples{
DNA <- Make_DNA(4)
DNA
}
