% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcription}
\alias{transcription}
\title{Transcription}
\usage{
transcription(dna_list)
}
\arguments{
\item{dna_list}{a list one-stringed DNA.}
}
\value{
a list of mRNA
}
\description{
Transcription
}
