% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translator.R
\name{translator}
\alias{translator}
\title{Translation}
\usage{
translator(codons)
}
\arguments{
\item{codons}{a triplet coding for an aminoacid}
}
\value{
reads codon and gives aminoacid
}
\description{
Translation
}
