% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{triplets}
\alias{triplets}
\title{Triplets}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
triplets
}
\description{
All triplets coding for all amino acids. 3 x bases --> 1 AA
}
\keyword{datasets}
