% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcription}
\alias{transcription}
\title{This function transcript a DNA sequence into an mRNA sequence}
\usage{
transcription(dna_seq)
}
\arguments{
\item{dna_seq}{should be a DNA sequence as the input of the function}
}
\value{
The function returns the mRNA sequence
}
\description{
This function transcript a DNA sequence into an mRNA sequence
}
\examples{
mRNA <- transcription("TCGGA")

}
